+++ src/pages/cart/cart.html
@@ -7,7 +7,6 @@
     <link rel="icon" href="./favicon.ico" />
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/variable/pretendardvariable.min.css" />
     <script type="module" src="/src/pages/cart/cart.js"></script>
-    <script type="module" src="/main.js"></script>
   </head>
   <body>
     <c-header></c-header>
